250322.19:41
первая альфа считай
+туалет
+новые спрайты
+новые скрипты
+фиксированное разрешение основанное на тайлах(32 на 32)
+механика РВОВ

+++ Assets/Скрипты/Здоровье/PlayerLives.cs
@@ -3,7 +3,7 @@
     public int Lives
     {
         get => 
-            PlayerLiveCounter.Active.LivesRemaining;
+            (int)PlayerLiveCounter.Active.LivesRemaining;
 
         set => 
             PlayerLiveCounter.Active.LivesRemaining = value;
